AVLSort

Description

Documentation

sortAVLAsc

Arguments

:: Ord a

=> [a]

-> [a]
sorted (ASC) input list

Function performs sort at list of elements of type a (instance of Ord) using AVLTree to maintain the sorting process. Each of N elements is inserting to the AVLTree in O(logN) time, so overally the time cost of building the tree is O(NlogN) time and O(N) space. After building the tree it is simply linearized in O(N) time. Sorting order: ASCENDING