AVLTree

Description

Documentation

data AVLTree a b

Definiction of AVL tree.

newTree

Arguments

:: AVLTree a b
Output: 'AVLTree a b' build with single EmptyNode

Function returns AVLTree as an EmptyNode

linearOrder

Arguments

:: Ord a

=> AVLTree a b
'AVLTree a b' to linearize

-> [b]
List of values in ascending order

Function returns list that consists of AVLTree values in ascending order of keys

reversedOrder

Arguments

:: Ord a

=> AVLTree a b
'AVLTree a b' to linearize

-> [b]
Output List

Function returns list that consists of AVLTree values in descending order of keys

linearKeysOrder

Arguments

:: Ord a

=> AVLTree a b
'AVLTree a b' to linearize

-> [a]
List of keys in ascending order

Function that returns list of keys in AVLTree in ascending order

reversedKeysOrder

Arguments

:: Ord a

=> AVLTree a b
'AVLTree a b' to linearize

-> [a]
Output List

Function returns list that consists of AVLTree keys in descending order of keys

insert

Arguments

:: Ord a

=> a
Key of element to be inserted

-> b
Value of the element to be inserted

-> AVLTree a b
'AVLTree a b' to which the element will be inserted

-> AVLTree a b
'AVLTree a b' after insertion of element

Function that insert an element key of type a and value of type b into the AVLTree a b. This version of insert function provides a straightforward insert i.e. each element may appears multiple times at the tree.

insertSingleton

Arguments

:: Ord a

=> a
Key of element to be inserted

-> b
Value of the element to be inserted

-> AVLTree a b
'AVLTree a b' to which the element will be inserted

-> AVLTree a b
'AVLTree a b' after insertion of element

Function that insert an element key of type a and value of type b into the AVLTree a b. This version of insert function provides a 'Set-like' insert i.e. each element may appears only once at the tree. Overall concept of insertion operation in functional language like Haskell comes from https://gist.github.com/timjb/8292342 - but it's not simply a copy of this code

insertKeyAsValue

Arguments

:: Ord a

=> a
Key and value of element to be inserted

-> AVLTree a ()
'AVLTree a ()' to which the element will be inserted

-> AVLTree a ()
'AVLTree a ()' after insertion of element

Function inserts the same value as key and value of AVLNode using for that insert function (see above). Repetition of keys in tree - allowed.

insertKeyAsValueSingleton

Arguments

:: Ord a

=> a
Key and value of element to be inserted

-> AVLTree a ()
'AVLTree a ()' to which the element will be inserted

-> AVLTree a ()
'AVLTree a ()' after insertion of element

Function inserts the same value as key and value of AVLNode using for that insert function (see above). Repetition of keys in tree - forbidden.

debugShow

Arguments

:: (Show a, Show b)

=> AVLTree a b
'AVLTree a b' to be printed in debug mode

-> String
String representation of tree

Debug function which provides a way to print additional info about balance coefficients of nodes

depth

Arguments

:: AVLTree a b
'AVLTree a b' to measure depth

-> Int
Int depth of tree (with regards that depth of EmptyNode = 0)

Function that returns depth of tree

containsKey

Arguments

:: Ord a

=> a
Key to be found of type a

-> AVLTree a b
'AVLTree a b' to find key in

-> Bool
result - indicates the success status of search

Function that determines whether a tree conatins given key or not

delete

Arguments

:: Ord a

=> a
key of type a to be deleted

-> AVLTree a b
'AVLTree a b' - the tree to operate on

-> (AVLTree a b, Maybe b)
output as described above

Function that performs delete operation at AVL Tree returns pair (value, AVLTree) where value is the value associated with deleted key and AVLTree is a tree after delete with height keeped O(logN) if the key is not in a tree the function returns Just

getValueOfKey

Arguments

:: Ord a

=> a
searching ket of type a

-> AVLTree a b
'AVLTree a b' to look for element

-> (AVLTree a b, Maybe b)
result as described above

Function that get element with a given key from given AVLTree without deleting it. The function returns (AVLTree, Maybe value) so that in case of success the requested value is placed in Just value, otherwise the second element of pair is Nothing