HashMap

Description

Documentation

newHashMap :: HashMap a b

Function that returns new empty HashMap

insertKeyVal

Arguments

:: (Eq a, Hashable a)

=> a
key

-> b
value

-> HashMap a b
'HashMap a b' to insert to

-> HashMap a b
'HashMap a b' after insertion

Function that inserts key and value into the given HashMap

getElement

Arguments

:: (Eq a, Hashable a)

=> a
key of type a

-> HashMap a b
'HashMap a b' to search to

-> (HashMap a b, Maybe b)
result as described above

Function that get element with a given key from given HashMap without deleting it. The function returns (HashMap, Maybe value) so that in case of success the requested value is placed in Just value, otherwise the second element of pair is Nothing

getKeys

Arguments

:: Eq a

=> HashMap a b
'HashMap a b' to take keys

-> [a]
set of keys

Function that returns key set from HashMap - the order is random

getValues

Arguments

:: Eq a

=> HashMap a b
'HashMap a b' to take keys

-> [b]
set of keys

Function that returns values set in given HashMap - the order is random

containsKey

Arguments

:: (Hashable a, Eq a)

=> a
Key of type a to check

-> HashMap a b
given 'HashMap a b'

-> Bool
result of check

Function that chechs whether HashMap contains given key

deleteFromHM

Arguments

:: (Hashable a, Eq a)

=> a
Key of type a to delete

-> HashMap a b
given 'HashMap a b'

-> (HashMap a b, Maybe b)
output as described above

Function deletes key and value from given HashMap and returns pair (HashMap, Maybe elem) - so in case of success - the HashMap after delete operation and deleted value, or otherwise pair (HashMap, Nothing) with unmodified HashMap