Stack

Description

Documentation

newtype Stack a

Definition of Stack

Constructors

Stack [a]

emptyStack

Arguments

:: Stack a
Function return (empty Stack)

Function that returns empty stack

isEmpty

Arguments

:: Stack a
Input 'Stack a'

-> Bool
Boolean value indicates emptiness of input 'Stack a'

Function that cheks whether tge given Stack is empty

toList

Arguments

:: Stack a
Input 'Stack a'

-> [a]
Output: list of elements from the stack in LIFO order

Function that produces list from given Stack in reversed insertion order

push

Arguments

:: a
Element of type a to be pushed on stack

-> Stack a
'Stack a' to push the element

-> Stack a
'Stack a' after the operation

Function that pushes element at the top of stack

pop

Arguments

:: Stack a
'Stack a' to be popped from

-> (Stack a, Maybe a)

Function that pops the top alement from stack returns this element in pair with the stack after the operation. If the stack is initially empty - the function returns Nothing